feat: 챌린지 상세 페이지 구조 변경
